Breaking http server out from the rest of the codebase.
