+++ bin/run
@@ -1,5 +1,5 @@
 #!/usr/bin/env node
 
-console.log('starting http configurable mock server')
+console.log('starting run-amock http server')
 
 import '../server.js'
