+++ lib/coreHandlers.js
@@ -74,10 +74,6 @@ const addMocksEndpoint = (request) => {
       })
     })
   })
-
-  if (isDebug) {
-    console.log('Configured handlers (after setup)', configuredHandlers)
-  }
   if (errors.length > 0) {
     console.error('Errors when setting up handlers', errors)
     return respondWith400(`There were errors handling your stub setup: ${errors.join(', ')}`)
